requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0

pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

python-dotenv>=1.0.0
requests>=2.31.0
langsmith>=0.1.100

pytest>=8.0.0
